Combine motionDetailPage with motionDetailScreen
